server/requirements.txt
auth==0.5.3
bcrypt==4.1.2
blinker==1.8.2
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.1.3
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
eventlet==0.37.0
falcon==3.1.3
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.53.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.1
matplotlib==3.9.0
mongoengine==0.29.1
mysql-connector-python==9.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pipenv==2024.0.3
platformdirs==4.3.6
pyasn1==0.6.1
pymongo==4.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
requests==2.27.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.12
virtualenv==20.26.5
Werkzeug==3.0.4
WTForms==3.1.2